image_uploader/requirements.txt
Django==4.2
djangorestframework==3.14.0
boto3==1.26.0
Pillow==9.5.0